Licensing stuff removed.
It is now possible to run Windows build with run target.

On Windows LZ installed with Install4J now fails with error.
As this does not occur when running with run target, it seems
to be Install4J related problem.

I plan to replace Install4J with IZpack, so it will be resolved with
the change.