Use metadata from embedded JPEG thumbnail
This is not the most elegant way to do this, but by far the easiest.
Better would be to lookup the IFD in the bytebuffer and put the
ExifMetadataReader to work. It seems to me that this involves a lot of
rework in the EXIF code, as it assumes all offsets to be absolute to the
beginning of the file, which is not the case when processing embedded
JPEGs.